Own module for Product Main seller
