+++ grap_change_views_mrp/readme/DESCRIPTION.rst
@@ -1,4 +1,1 @@
-This module only install other modules, that are required in R&D for using
-MRP in Odoo.
-
-In that case, just add ``-i grap_mrp``.
+Change default views for ``mrp`` modules.
