+++ product_main_seller/README.rst
@@ -1,0 +1,63 @@
+==========================
+GRAP - Product Main Seller
+==========================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-quentinDupont%2Fgrap--odoo--custom-lightgray.png?logo=github
+    :target: https://github.com/quentinDupont/grap-odoo-custom/tree/12.0_MRP_Grap/product_main_seller
+    :alt: quentinDupont/grap-odoo-custom
+
+|badge1| |badge2| |badge3| 
+
+Get main seller partner for a product  
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Known issues / Roadmap
+======================
+
+* Main seller partner could be computed according to purchases statistics
+* Company parameter to choose between differents ways to know which seller is the main one
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/quentinDupont/grap-odoo-custom/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/quentinDupont/grap-odoo-custom/issues/new?body=module:%20product_main_seller%0Aversion:%2012.0_MRP_Grap%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* GRAP
+
+Contributors
+~~~~~~~~~~~~
+
+* Quentin Dupont (quentin.dupont@grap.coop)
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `quentinDupont/grap-odoo-custom <https://github.com/quentinDupont/grap-odoo-custom/tree/12.0_MRP_Grap/product_main_seller>`_ project on GitHub.
+
+You are welcome to contribute.
