+++ product_main_seller/__manifest__.py
@@ -1,0 +1,17 @@
+# Copyright (C) 2021 - Today: GRAP (http://www.grap.coop)
+# @author: Quentin Dupont (quentin.dupont@grap.coop)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+{
+    "name": "GRAP - Product Main Seller",
+    "summary": "Product Attribute - Main seller for a product",
+    "version": "12.0.1.1.0",
+    "category": "GRAP - Custom",
+    "author": "GRAP",
+    "website": "https://github.com/grap/grap-odoo-custom",
+    "license": "AGPL-3",
+    "depends": [
+        "product",
+    ],
+    "installable": True,
+}
