+++ product_main_seller/models/__init__.py
@@ -1,0 +1,1 @@
+from . import product_product
