+++ product_main_seller/readme/CONTRIBUTORS.rst
@@ -1,0 +1,1 @@
+* Quentin Dupont (quentin.dupont@grap.coop)
