+++ product_main_seller/readme/DESCRIPTION.rst
@@ -1,0 +1,1 @@
+Get main seller partner for a product
