+++ product_main_seller/readme/ROADMAP.rst
@@ -1,0 +1,2 @@
+* Main seller partner could be computed according to purchases statistics
+* Company parameter to choose between differents ways to know which seller is the main one
